template

+++ cf_template/io.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 using i64 = long long;
 using u64 = unsigned long long;
-#define REP(i, n) for (int i = 0; (i64)(i) < (i64)(n); ++i)
+#define REP(i, n) for (int i = 0; i < (i64)(n); ++i)
 #define WHOLE(x) std::begin(x), std::end(x)
 #define LEN(a) int((a).size())
